server: move balance schedulers into schedulers directory. (#729)
