+++ server/balancer_test.go
@@ -20,8 +20,10 @@ import (
 	. "github.com/pingcap/check"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/kvproto/pkg/pdpb"
+	"github.com/pingcap/pd/server/cache"
 	"github.com/pingcap/pd/server/core"
 	"github.com/pingcap/pd/server/schedule"
+	"github.com/pingcap/pd/server/schedulers"
 )
 
 type testClusterInfo struct {
@@ -262,14 +264,14 @@ var _ = Suite(&testBalanceLeaderSchedulerSuite{})
 type testBalanceLeaderSchedulerSuite struct {
 	cluster *clusterInfo
 	tc      *testClusterInfo
-	lb      *balanceLeaderScheduler
+	lb      schedule.Scheduler
 }
 
 func (s *testBalanceLeaderSchedulerSuite) SetUpTest(c *C) {
 	s.cluster = newClusterInfo(newMockIDAllocator())
 	s.tc = newTestClusterInfo(s.cluster)
 	_, opt := newTestScheduleConfig()
-	s.lb = newBalanceLeaderScheduler(opt)
+	s.lb = schedulers.NewBalanceLeaderScheduler(opt)
 }
 
 func (s *testBalanceLeaderSchedulerSuite) schedule() schedule.Operator {
@@ -385,12 +387,23 @@ var _ = Suite(&testBalanceRegionSchedulerSuite{})
 
 type testBalanceRegionSchedulerSuite struct{}
 
+// TODO: remove it after moving tests to schedulers directory.
+func (s *testBalanceRegionSchedulerSuite) getCache(scheduler schedule.Scheduler) *cache.TTLUint64 {
+	type hasCache interface {
+		GetCache() *cache.TTLUint64
+	}
+	if c, ok := scheduler.(hasCache); ok {
+		return c.GetCache()
+	}
+	return nil
+}
+
 func (s *testBalanceRegionSchedulerSuite) TestBalance(c *C) {
 	cluster := newClusterInfo(newMockIDAllocator())
 	tc := newTestClusterInfo(cluster)
 
 	_, opt := newTestScheduleConfig()
-	sb := newBalanceRegionScheduler(opt)
+	sb := schedulers.NewBalanceRegionScheduler(opt)
 
 	opt.SetMaxReplicas(1)
 
@@ -409,7 +422,7 @@ func (s *testBalanceRegionSchedulerSuite) TestBalance(c *C) {
 	c.Assert(sb.Schedule(cluster), IsNil)
 	// 9 - 6 >= 2
 	tc.updateRegionCount(2, 6)
-	sb.cache.Remove(4)
+	s.getCache(sb).Remove(4)
 	// When store 1 is offline, it will be filtered,
 	// store 2 becomes the store with least regions.
 	checkTransferPeer(c, sb.Schedule(cluster), 4, 2)
@@ -428,7 +441,7 @@ func (s *testBalanceRegionSchedulerSuite) TestReplicas3(c *C) {
 	_, opt := newTestScheduleConfig()
 	opt.rep = newTestReplication(3, "zone", "rack", "host")
 
-	sb := newBalanceRegionScheduler(opt)
+	sb := schedulers.NewBalanceRegionScheduler(opt)
 
 	// Store 1 has the largest region score, so the balancer try to replace peer in store 1.
 	tc.addLabelsStore(1, 6, map[string]string{"zone": "z1", "rack": "r1", "host": "h1"})
@@ -439,7 +452,7 @@ func (s *testBalanceRegionSchedulerSuite) TestReplicas3(c *C) {
 	// This schedule try to replace peer in store 1, but we have no other stores,
 	// so store 1 will be set in the cache and skipped next schedule.
 	c.Assert(sb.Schedule(cluster), IsNil)
-	c.Assert(sb.cache.Exists(1), IsTrue)
+	c.Assert(s.getCache(sb).Exists(1), IsTrue)
 
 	// Store 4 has smaller region score than store 2.
 	tc.addLabelsStore(4, 2, map[string]string{"zone": "z1", "rack": "r2", "host": "h1"})
@@ -447,7 +460,7 @@ func (s *testBalanceRegionSchedulerSuite) TestReplicas3(c *C) {
 
 	// Store 5 has smaller region score than store 1.
 	tc.addLabelsStore(5, 2, map[string]string{"zone": "z1", "rack": "r1", "host": "h1"})
-	sb.cache.Remove(1) // Delete store 1 from cache, or it will be skipped.
+	s.getCache(sb).Remove(1) // Delete store 1 from cache, or it will be skipped.
 	checkTransferPeer(c, sb.Schedule(cluster), 1, 5)
 
 	// Store 6 has smaller region score than store 5.
@@ -461,10 +474,10 @@ func (s *testBalanceRegionSchedulerSuite) TestReplicas3(c *C) {
 	// If store 7 is not available, we wait.
 	tc.setStoreDown(7)
 	c.Assert(sb.Schedule(cluster), IsNil)
-	c.Assert(sb.cache.Exists(1), IsTrue)
+	c.Assert(s.getCache(sb).Exists(1), IsTrue)
 	tc.setStoreUp(7)
 	checkTransferPeer(c, sb.Schedule(cluster), 2, 7)
-	sb.cache.Remove(1)
+	s.getCache(sb).Remove(1)
 	checkTransferPeer(c, sb.Schedule(cluster), 1, 7)
 
 	// Store 8 has smaller region score than store 7, but the distinct score decrease.
@@ -477,8 +490,8 @@ func (s *testBalanceRegionSchedulerSuite) TestReplicas3(c *C) {
 	tc.setStoreDown(6)
 	tc.setStoreDown(7)
 	c.Assert(sb.Schedule(cluster), IsNil)
-	c.Assert(sb.cache.Exists(1), IsTrue)
-	sb.cache.Remove(1)
+	c.Assert(s.getCache(sb).Exists(1), IsTrue)
+	s.getCache(sb).Remove(1)
 
 	// Store 9 has different zone with other stores but larger region score than store 1.
 	tc.addLabelsStore(9, 9, map[string]string{"zone": "z2", "rack": "r1", "host": "h1"})
@@ -492,7 +505,7 @@ func (s *testBalanceRegionSchedulerSuite) TestReplicas5(c *C) {
 	_, opt := newTestScheduleConfig()
 	opt.rep = newTestReplication(5, "zone", "rack", "host")
 
-	sb := newBalanceRegionScheduler(opt)
+	sb := schedulers.NewBalanceRegionScheduler(opt)
 
 	tc.addLabelsStore(1, 4, map[string]string{"zone": "z1", "rack": "r1", "host": "h1"})
 	tc.addLabelsStore(2, 5, map[string]string{"zone": "z2", "rack": "r1", "host": "h1"})
@@ -527,7 +540,7 @@ func (s *testBalanceRegionSchedulerSuite) TestStoreWeight(c *C) {
 	tc := newTestClusterInfo(cluster)
 
 	_, opt := newTestScheduleConfig()
-	sb := newBalanceRegionScheduler(opt)
+	sb := schedulers.NewBalanceRegionScheduler(opt)
 	opt.SetMaxReplicas(1)
 
 	tc.addRegionStore(1, 10)
