+++ server/schedulers/utils.go
@@ -15,8 +15,10 @@ package schedulers
 
 import (
 	"math"
+	"time"
 
-	"github.com/ngaut/log"
+	log "github.com/Sirupsen/logrus"
+	"github.com/montanaflynn/stats"
 	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/pd/server/core"
 	"github.com/pingcap/pd/server/schedule"
@@ -114,4 +116,66 @@ func scheduleAddPeer(cluster schedule.Cluster, s schedule.Selector, filters ...s
 	}
 
 	return newPeer
+}
+
+func minUint64(a, b uint64) uint64 {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+func maxUint64(a, b uint64) uint64 {
+	if a > b {
+		return a
+	}
+	return b
+}
+
+func minDuration(a, b time.Duration) time.Duration {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+const (
+	bootstrapBalanceCount = 10
+	bootstrapBalanceDiff  = 2
+)
+
+// minBalanceDiff returns the minimal diff to do balance. The formula is based
+// on experience to let the diff increase alone with the count slowly.
+func minBalanceDiff(count uint64) float64 {
+	if count < bootstrapBalanceCount {
+		return bootstrapBalanceDiff
+	}
+	return math.Sqrt(float64(count))
+}
+
+// shouldBalance returns true if we should balance the source and target store.
+// The min balance diff provides a buffer to make the cluster stable, so that we
+// don't need to schedule very frequently.
+func shouldBalance(source, target *core.StoreInfo, kind core.ResourceKind) bool {
+	sourceCount := source.ResourceCount(kind)
+	sourceScore := source.ResourceScore(kind)
+	targetScore := target.ResourceScore(kind)
+	if targetScore >= sourceScore {
+		return false
+	}
+	diffRatio := 1 - targetScore/sourceScore
+	diffCount := diffRatio * float64(sourceCount)
+	return diffCount >= minBalanceDiff(sourceCount)
+}
+
+func adjustBalanceLimit(cluster schedule.Cluster, kind core.ResourceKind) uint64 {
+	stores := cluster.GetStores()
+	counts := make([]float64, 0, len(stores))
+	for _, s := range stores {
+		if s.IsUp() {
+			counts = append(counts, float64(s.ResourceCount(kind)))
+		}
+	}
+	limit, _ := stats.StandardDeviation(stats.Float64Data(counts))
+	return maxUint64(1, uint64(limit))
 }
