+++ server/util_test.go
@@ -20,28 +20,6 @@ import (
 	. "github.com/pingcap/check"
 )
 
-var _ = Suite(&testMinMaxSuite{})
-
-type testMinMaxSuite struct{}
-
-func (s *testMinMaxSuite) TestMinUint64(c *C) {
-	c.Assert(minUint64(1, 2), Equals, uint64(1))
-	c.Assert(minUint64(2, 1), Equals, uint64(1))
-	c.Assert(minUint64(1, 1), Equals, uint64(1))
-}
-
-func (s *testMinMaxSuite) TestMaxUint64(c *C) {
-	c.Assert(maxUint64(1, 2), Equals, uint64(2))
-	c.Assert(maxUint64(2, 1), Equals, uint64(2))
-	c.Assert(maxUint64(1, 1), Equals, uint64(1))
-}
-
-func (s *testMinMaxSuite) TestMinDuration(c *C) {
-	c.Assert(minDuration(time.Minute, time.Second), Equals, time.Second)
-	c.Assert(minDuration(time.Second, time.Minute), Equals, time.Second)
-	c.Assert(minDuration(time.Second, time.Second), Equals, time.Second)
-}
-
 var _ = Suite(&testUtilSuite{})
 
 type testUtilSuite struct{}
